% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_compress.R
\name{compress}
\alias{compress}
\title{Compress images}
\usage{
compress(img, b)
}
\arguments{
\item{img}{array}

\item{b}{integer}
}
\value{
array
}
\description{
Compress the size an image by reducing the number of bits
used to represent each pixel to b
}
\examples{
compress(img,4)
}
