% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_compression.R
\name{compression}
\alias{compression}
\title{Title}
\usage{
compression(img_path, b, out_path)
}
\arguments{
\item{img_path}{String , file path of the image}

\item{b}{Integer, desired byte size}

\item{out_path}{String, a valid output path depending on the OS}
}
\value{
String, compressed image path
}
\description{
Title
}
\examples{
# compression("../data/image.png", 2L, "..data/comrpessed_image.png")

}
