% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_crop.R
\name{crop}
\alias{crop}
\title{Crop images}
\usage{
crop(img_path, H, W, out_path)
}
\arguments{
\item{img_path}{---- String , file path of the image .}

\item{H}{---- Integer, the desired height of the cropped image}

\item{W}{---- Integer, the desired width of the cropped image}

\item{out_path}{---- String , desired file path of the cropped image}
}
\value{
---String , cropped image path and saves the image .
}
\description{
Crop images
}
\examples{


}
