% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_size.R
\name{image_size}
\alias{image_size}
\title{Get image size}
\usage{
image_size(img)
}
\arguments{
\item{img}{String . path of image}
}
\value{
integer
}
\description{
Calculate the size of an image
}
\examples{
# image_size("../data/image.png")
}
