% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageCompressR-package.R
\docType{package}
\name{imageCompressR-package}
\alias{imageCompressR}
\alias{imageCompressR-package}
\title{imageCompressR: It compresses images}
\description{
It compresses image shape by cropping and reducing memory size .
}
\author{
\strong{Maintainer}: Sayanti,Aditya,Alden Ghosh,Sharma,Chen \email{first.last@example.com}

}
\keyword{internal}
