% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_seam.R
\name{seam_carve}
\alias{seam_carve}
\title{Title It resizes the image by removing lowest energy seam veritically and horizontally}
\usage{
seam_carve(img, h, w)
}
\arguments{
\item{img}{array}

\item{h}{integer}

\item{w}{integer}
}
\value{
array
}
\description{
Title It resizes the image by removing lowest energy seam veritically and horizontally
}
\examples{
seam_carve(img,4)
}
