% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_size.R
\name{size}
\alias{size}
\title{Title Calculates the memory size of image}
\usage{
size(img, b)
}
\arguments{
\item{img}{array}

\item{b}{interger}
}
\value{
integer
}
\description{
Title Calculates the memory size of image
}
\examples{
size(img,4)
}
