% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_compress.R
\name{compress}
\alias{compress}
\title{Compress images}
\usage{
compress(img_path, b, out_path)
}
\arguments{
\item{img_path}{String , file path of the image}

\item{b}{Integer, desired byte size}

\item{out_path}{String, a valid output path depending on the OS}
}
\value{
String, compressed image path
}
\description{
Compress the size an image by reducing the number of bits
used to represent each pixel to b
}
\examples{

}
