% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_compress.R
\name{compress}
\alias{compress}
\title{Title Compress images by memory size .}
\usage{
compress(img, b)
}
\arguments{
\item{img}{array}

\item{b}{integer}
}
\value{
array
}
\description{
Title Compress images by memory size .
}
\examples{
compress(img,4)
}
