% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageCompressR-package.R
\docType{package}
\name{imageCompressR-package}
\alias{imageCompressR}
\alias{imageCompressR-package}
\title{imageCompressR: It compresses images}
\description{
It compresses image shape by cropping ,seam carving and reducing memory size .
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
