% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageCompressR-package.R
\docType{package}
\name{imageCompressR-package}
\alias{imageCompressR}
\alias{imageCompressR-package}
\title{imageCompressR: It compresses images}
\description{
It compresses image shape by cropping and reducing memory size.
    Compress function reduces the byte size of image, 
    crop function crops the image form all sides.
    Image_size function measures the size of the image (memory).
}
\author{
\strong{Maintainer}: Sayanti Ghosh \email{sayanti.1986.ghosh@gmail.com}

Authors:
\itemize{
  \item Aditya sharma \email{sayanti.1986.ghosh@gmail.com}
  \item Alden Chen \email{sayanti.1986.ghosh@gmail.com}
}

}
\keyword{internal}
